update(admin): products-delete

+++ admin/products/index.php
@@ -56,6 +56,7 @@
                             <th scope="col">Kích thước</th>
                             <th scope="col">Giá</th>
                             <th scope="col">Thể loại</th>
+                            <th scope="col">Trạng thái</th>
                             <th scope="col">Người thêm</th>
                             <th scope="col">Sửa</th>
                             <th scope="col">Xóa</th>
@@ -74,17 +75,32 @@
                                     <td><?= $each['size'] ?> cm</td>
                                     <td><?= number_format($each['price'], 0, '.', ' ') ?>&#8363</td>
                                     <td><?= $each['category_name'] ?></td>
+                                    <td><?php 
+                                        if($each['status'] === '1') { ?>
+                                            Đang bán
+                                        <?php } else { ?>
+                                            Ngừng bán
+                                        <?php } ?>
+                                    </td>
                                     <td><?= $each['admin_name'] ?></td>
-                                    <td>
-                                        <a href="form_update.php?id=<?= $each['id'] ?>&admin_id=<?= $each['admin_id'] ?>">
-                                            <i class="bi bi-pencil-fill"></i>
-                                        </a>
-                                    </td>
-                                    <td>
-                                        <a href="delete.php?id=<?= $each['id'] ?>&admin_id=<?= $each['admin_id'] ?>">
-                                        <i class="bi bi-trash-fill"></i>
-                                        </a>
-                                    </td>
+                                    <?php if($each['status'] === '1') { ?>
+                                        <td>
+                                            <a href="form_update.php?id=<?= $each['id'] ?>&admin_id=<?= $each['admin_id'] ?>">
+                                                <i class="bi bi-pencil-fill"></i>
+                                            </a>
+                                        </td>
+                                        <td>
+                                            <a href="delete.php?id=<?= $each['id'] ?>&admin_id=<?= $each['admin_id'] ?>">
+                                            <i class="bi bi-trash-fill"></i>
+                                            </a>
+                                        </td>
+                                    <?php } else { ?>
+                                        <td colspan="2">
+                                            <a href="update_status.php?id=<?= $each['id'] ?>&admin_id=<?= $each['admin_id'] ?>">
+                                                Mở bán
+                                            </a>
+                                        </td>
+                                    <?php } ?>
                                 </tr>
                             <?php } ?>
